+++ core/src/main/java/org/infinispan/interceptors/impl/AsyncInterceptorChainImpl.java
@@ -55,7 +55,7 @@ public class AsyncInterceptorChainImpl implements AsyncInterceptorChain {
    @Start
    private void printChainInfo() {
       if (log.isDebugEnabled()) {
-         log.debugf("Interceptor chain size: %d", size());
+         log.debugf("Interceptor chain size: %d", (Integer)size());
          log.debugf("Interceptor chain is: %s", toString());
       }
    }
